Abstract ModelStub into its own class
Code duplication is occuring due to ModelStub being implemented in
all three add commands individually.

We should move the ModelStub into its own class to reduce the
amount of duplicated code.
