+++ src/test/java/seedu/address/logic/commands/AddLocationCommandTest.java
@@ -18,6 +18,7 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.LocationBook;
 import seedu.address.model.Model;
+import seedu.address.model.ModelStub;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyLocationBook;
 import seedu.address.model.ReadOnlyUserPrefs;
@@ -50,7 +51,7 @@ public class AddLocationCommandTest {
     public void execute_duplicateLocation_throwsCommandException() {
         Location validLocation = new LocationBuilder().build();
         AddLocationCommand addLocationCommand = new AddLocationCommand(validLocation);
-        AddLocationCommandTest.ModelStub modelStub = new AddLocationCommandTest.ModelStubWithLocation(validLocation);
+        ModelStub modelStub = new AddLocationCommandTest.ModelStubWithLocation(validLocation);
 
         assertThrows(CommandException.class, AddLocationCommand.MESSAGE_DUPLICATE_LOCATION, () ->
                 addLocationCommand.execute(modelStub));
@@ -81,169 +82,9 @@ public class AddLocationCommandTest {
     }
 
     /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook addressBook) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePerson(Person target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPerson(Person target, Person editedPerson) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Location> getFilteredLocationList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPersonList(Predicate<Person> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredLocationList(Predicate<Location> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getLocationBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setLocationBookFilePath(Path locationBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addLocation(Location location) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasVisit(Visit visit) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteVisit(Visit visit) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addVisit(Visit visit) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getVisitBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setVisitBookFilePath(Path visitBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setVisitBook(ReadOnlyVisitBook visitBook) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyVisitBook getVisitBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Visit> getFilteredVisitList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredVisitList(Predicate<Visit> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setLocationBook(ReadOnlyLocationBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyLocationBook getLocationBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasLocation(Location location) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
      * A Model stub that contains a single location.
      */
-    private class ModelStubWithLocation extends AddLocationCommandTest.ModelStub {
+    private class ModelStubWithLocation extends ModelStub {
         private final Location location;
 
         ModelStubWithLocation(Location location) {
@@ -261,7 +102,7 @@ public class AddLocationCommandTest {
     /**
      * A Model stub that always accept the location being added.
      */
-    private class ModelStubAcceptingLocationAdded extends AddLocationCommandTest.ModelStub {
+    private class ModelStubAcceptingLocationAdded extends ModelStub {
         final ArrayList<Location> locationsAdded = new ArrayList<>();
 
         @Override
