+++ src/test/java/seedu/address/logic/commands/AddVisitCommandTest.java
@@ -17,6 +17,7 @@ import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
+import seedu.address.model.ModelStub;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyLocationBook;
 import seedu.address.model.ReadOnlyUserPrefs;
@@ -50,7 +51,7 @@ public class AddVisitCommandTest {
     public void execute_duplicateVisit_throwsCommandException() {
         Visit validVisit = new VisitBuilder().build();
         AddVisitCommand addvisitCommand = new AddVisitCommand(validVisit);
-        AddVisitCommandTest.ModelStub modelStub = new AddVisitCommandTest.ModelStubWithVisit(validVisit);
+        ModelStub modelStub = new AddVisitCommandTest.ModelStubWithVisit(validVisit);
 
         assertThrows(CommandException.class, AddVisitCommand.MESSAGE_DUPLICATE_VISIT, () ->
                 addvisitCommand.execute(modelStub));
@@ -81,168 +82,9 @@ public class AddVisitCommandTest {
     }
 
     /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook addressBook) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePerson(Person target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPerson(Person target, Person editedPerson) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPersonList(Predicate<Person> predicate) {
-
-        }
-
-        @Override
-        public boolean hasLocation(Location location) {
-            return false;
-        }
-
-        @Override
-        public Path getLocationBookFilePath() {
-            return null;
-        }
-
-        @Override
-        public void setLocationBookFilePath(Path locationBookFilePath) {
-
-        }
-
-        @Override
-        public void setLocationBook(ReadOnlyLocationBook locationBook) {
-
-        }
-
-        @Override
-        public ReadOnlyLocationBook getLocationBook() {
-            return null;
-        }
-
-        @Override
-        public void addLocation(Location location) {
-
-        }
-
-        @Override
-        public ObservableList<Location> getFilteredLocationList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        public void updateFilteredLocationList(Predicate<Location> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getVisitBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setVisitBookFilePath(Path visitBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addVisit(Visit visit) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteVisit(Visit visit) {
-
-        }
-
-        @Override
-        public void setVisitBook(ReadOnlyVisitBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyVisitBook getVisitBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Visit> getFilteredVisitList() {
-            return null;
-        }
-
-        @Override
-        public void updateFilteredVisitList(Predicate<Visit> predicate) {
-
-        }
-
-        @Override
-        public boolean hasVisit(Visit visit) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
      * A Model stub that contains a single visit.
      */
-    private class ModelStubWithVisit extends AddVisitCommandTest.ModelStub {
+    private class ModelStubWithVisit extends ModelStub {
         private final Visit visit;
 
         ModelStubWithVisit(Visit visit) {
@@ -260,7 +102,7 @@ public class AddVisitCommandTest {
     /**
      * A Model stub that always accept the visit being added.
      */
-    private class ModelStubAcceptingVisitAdded extends AddVisitCommandTest.ModelStub {
+    private class ModelStubAcceptingVisitAdded extends ModelStub {
         final ArrayList<Visit> visitsAdded = new ArrayList<>();
 
         @Override
